fix(compiler): keep artifact fingerprints core-only
